feat: enhance attestation process with optional TCB status retrieval
- Updated README to clarify the optional nature of TCB_STATUS and added VERIFIER_ENDPOINT for enclave status checks.
- Modified attestation request validation to utilize whitelisted domains from configuration.
- Implemented TCB status retrieval from a verifier endpoint, allowing for dynamic enclave status evaluation.
- Introduced error handling for TCB status retrieval in configuration validation.
- Added new error type for decoding quote failures in the error handling module.

+++ internal/services/enclave_info/enclave_info.go
@@ -209,7 +209,7 @@ func createSgxInfoFromReportData(reportData *SGXReportData) SgxInfo {
 		SecurityVersion: binary.LittleEndian.Uint16(reportData.SecurityVersion),
 		Debug:           reportData.Debug,
 		Aleo:            aleoInfo,
-		TCBStatus:       uint(configs.GetAppConfig().TCBStatus),
+		TCBStatus:       configs.GetAppConfig().TCBStatus,
 	}
 
 	return sgxInfo
